fix(task-delete): resolve bearer auth

+++ src/api/project/task/task.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Request } from 'express';
@@ -26,9 +26,16 @@ export class TaskService {
     return await this.repository.findOne({where: {id}, relations: ['project']});
   }
 
-  async deleteOneById(id: number): Promise<void> {
+  async deleteOneById(id: number, user: User): Promise<HttpException> {
     const task = await this.repository.findOne({ where: { id }, relations: ['project'] });
+
+    if (!task) {
+      throw new NotFoundException('Tâche introuvable');
+    }
+    
     await this.repository.delete(task.id)
+
+    return new HttpException('Tâche supprimée', HttpStatus.OK);
   }
 
   public async getAllFromProject(project: number): Promise<Task[]>{
